Refactor project to .NetStandard 2.0 and .NetFramework 4.6.2
Updated projects to more recent versions of language and support cross-platform development with NetStandard

Removed libraries included with repo and replaced dependencies with NuGet includes.

Refactored names of projects to be consistent across solution
