+++ DotNetAdapterSDK.sln
@@ -1,54 +1,65 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 2012
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DotNetAdapterSDK", "DotNetAdapterSDK.csproj", "{D8D08DBE-B511-4DE0-B5A9-563EBFC60C4F}"
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.29306.81
+MinimumVisualStudioVersion = 15.0.26124.0
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "mtconnect_adapter_sdk", "src\mtconnect_adapter_sdk.csproj", "{80FB2820-86E2-4A87-9565-ED0227E54E7A}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NUnitTests", "NUnitTests\NUnitTests.csproj", "{429B8B5C-8483-4680-BC8D-E2792574F6A4}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "mtconnect_adapter_sdk.utests", "test\mtconnect_adapter_sdk.utests.csproj", "{91514352-32BE-469F-ACE2-12AE559CF162}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AdapterLab", "Sample\AdapterLab.csproj", "{AD661DEE-4205-4289-9E77-E4F264809DE3}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AdapterLab", "sample\AdapterLab.csproj", "{AD8454F6-F317-4279-A2E0-E0DCC7AF5B2A}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
-		Debug|Mixed Platforms = Debug|Mixed Platforms
+		Debug|x64 = Debug|x64
 		Debug|x86 = Debug|x86
 		Release|Any CPU = Release|Any CPU
-		Release|Mixed Platforms = Release|Mixed Platforms
+		Release|x64 = Release|x64
 		Release|x86 = Release|x86
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{D8D08DBE-B511-4DE0-B5A9-563EBFC60C4F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{D8D08DBE-B511-4DE0-B5A9-563EBFC60C4F}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{D8D08DBE-B511-4DE0-B5A9-563EBFC60C4F}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
-		{D8D08DBE-B511-4DE0-B5A9-563EBFC60C4F}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
-		{D8D08DBE-B511-4DE0-B5A9-563EBFC60C4F}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{D8D08DBE-B511-4DE0-B5A9-563EBFC60C4F}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{D8D08DBE-B511-4DE0-B5A9-563EBFC60C4F}.Release|Any CPU.Build.0 = Release|Any CPU
-		{D8D08DBE-B511-4DE0-B5A9-563EBFC60C4F}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
-		{D8D08DBE-B511-4DE0-B5A9-563EBFC60C4F}.Release|Mixed Platforms.Build.0 = Release|Any CPU
-		{D8D08DBE-B511-4DE0-B5A9-563EBFC60C4F}.Release|x86.ActiveCfg = Release|Any CPU
-		{429B8B5C-8483-4680-BC8D-E2792574F6A4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{429B8B5C-8483-4680-BC8D-E2792574F6A4}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{429B8B5C-8483-4680-BC8D-E2792574F6A4}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
-		{429B8B5C-8483-4680-BC8D-E2792574F6A4}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
-		{429B8B5C-8483-4680-BC8D-E2792574F6A4}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{429B8B5C-8483-4680-BC8D-E2792574F6A4}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{429B8B5C-8483-4680-BC8D-E2792574F6A4}.Release|Any CPU.Build.0 = Release|Any CPU
-		{429B8B5C-8483-4680-BC8D-E2792574F6A4}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
-		{429B8B5C-8483-4680-BC8D-E2792574F6A4}.Release|Mixed Platforms.Build.0 = Release|Any CPU
-		{429B8B5C-8483-4680-BC8D-E2792574F6A4}.Release|x86.ActiveCfg = Release|Any CPU
-		{AD661DEE-4205-4289-9E77-E4F264809DE3}.Debug|Any CPU.ActiveCfg = Debug|x86
-		{AD661DEE-4205-4289-9E77-E4F264809DE3}.Debug|Mixed Platforms.ActiveCfg = Debug|x86
-		{AD661DEE-4205-4289-9E77-E4F264809DE3}.Debug|Mixed Platforms.Build.0 = Debug|x86
-		{AD661DEE-4205-4289-9E77-E4F264809DE3}.Debug|x86.ActiveCfg = Debug|x86
-		{AD661DEE-4205-4289-9E77-E4F264809DE3}.Debug|x86.Build.0 = Debug|x86
-		{AD661DEE-4205-4289-9E77-E4F264809DE3}.Release|Any CPU.ActiveCfg = Release|x86
-		{AD661DEE-4205-4289-9E77-E4F264809DE3}.Release|Mixed Platforms.ActiveCfg = Release|x86
-		{AD661DEE-4205-4289-9E77-E4F264809DE3}.Release|Mixed Platforms.Build.0 = Release|x86
-		{AD661DEE-4205-4289-9E77-E4F264809DE3}.Release|x86.ActiveCfg = Release|x86
-		{AD661DEE-4205-4289-9E77-E4F264809DE3}.Release|x86.Build.0 = Release|x86
+		{80FB2820-86E2-4A87-9565-ED0227E54E7A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{80FB2820-86E2-4A87-9565-ED0227E54E7A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{80FB2820-86E2-4A87-9565-ED0227E54E7A}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{80FB2820-86E2-4A87-9565-ED0227E54E7A}.Debug|x64.Build.0 = Debug|Any CPU
+		{80FB2820-86E2-4A87-9565-ED0227E54E7A}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{80FB2820-86E2-4A87-9565-ED0227E54E7A}.Debug|x86.Build.0 = Debug|Any CPU
+		{80FB2820-86E2-4A87-9565-ED0227E54E7A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{80FB2820-86E2-4A87-9565-ED0227E54E7A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{80FB2820-86E2-4A87-9565-ED0227E54E7A}.Release|x64.ActiveCfg = Release|Any CPU
+		{80FB2820-86E2-4A87-9565-ED0227E54E7A}.Release|x64.Build.0 = Release|Any CPU
+		{80FB2820-86E2-4A87-9565-ED0227E54E7A}.Release|x86.ActiveCfg = Release|Any CPU
+		{80FB2820-86E2-4A87-9565-ED0227E54E7A}.Release|x86.Build.0 = Release|Any CPU
+		{91514352-32BE-469F-ACE2-12AE559CF162}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{91514352-32BE-469F-ACE2-12AE559CF162}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{91514352-32BE-469F-ACE2-12AE559CF162}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{91514352-32BE-469F-ACE2-12AE559CF162}.Debug|x64.Build.0 = Debug|Any CPU
+		{91514352-32BE-469F-ACE2-12AE559CF162}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{91514352-32BE-469F-ACE2-12AE559CF162}.Debug|x86.Build.0 = Debug|Any CPU
+		{91514352-32BE-469F-ACE2-12AE559CF162}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{91514352-32BE-469F-ACE2-12AE559CF162}.Release|Any CPU.Build.0 = Release|Any CPU
+		{91514352-32BE-469F-ACE2-12AE559CF162}.Release|x64.ActiveCfg = Release|Any CPU
+		{91514352-32BE-469F-ACE2-12AE559CF162}.Release|x64.Build.0 = Release|Any CPU
+		{91514352-32BE-469F-ACE2-12AE559CF162}.Release|x86.ActiveCfg = Release|Any CPU
+		{91514352-32BE-469F-ACE2-12AE559CF162}.Release|x86.Build.0 = Release|Any CPU
+		{AD8454F6-F317-4279-A2E0-E0DCC7AF5B2A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{AD8454F6-F317-4279-A2E0-E0DCC7AF5B2A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{AD8454F6-F317-4279-A2E0-E0DCC7AF5B2A}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{AD8454F6-F317-4279-A2E0-E0DCC7AF5B2A}.Debug|x64.Build.0 = Debug|Any CPU
+		{AD8454F6-F317-4279-A2E0-E0DCC7AF5B2A}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{AD8454F6-F317-4279-A2E0-E0DCC7AF5B2A}.Debug|x86.Build.0 = Debug|Any CPU
+		{AD8454F6-F317-4279-A2E0-E0DCC7AF5B2A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{AD8454F6-F317-4279-A2E0-E0DCC7AF5B2A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{AD8454F6-F317-4279-A2E0-E0DCC7AF5B2A}.Release|x64.ActiveCfg = Release|Any CPU
+		{AD8454F6-F317-4279-A2E0-E0DCC7AF5B2A}.Release|x64.Build.0 = Release|Any CPU
+		{AD8454F6-F317-4279-A2E0-E0DCC7AF5B2A}.Release|x86.ActiveCfg = Release|Any CPU
+		{AD8454F6-F317-4279-A2E0-E0DCC7AF5B2A}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {87255246-4E86-4433-81CE-2C6B382A2005}
 	EndGlobalSection
 EndGlobal
