--- Sample/AdapterLab.sln
@@ -1,52 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 11.00
-# Visual Studio 2010
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AdapterLab", "AdapterLab.csproj", "{AD661DEE-4205-4289-9E77-E4F264809DE3}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{1915FA01-5C69-48FC-8D00-CE0EAF897177}"
-	ProjectSection(SolutionItems) = preProject
-		AdapterLab.vsmdi = AdapterLab.vsmdi
-		Local.testsettings = Local.testsettings
-		TraceAndTestImpact.testsettings = TraceAndTestImpact.testsettings
-	EndProjectSection
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NUnitTests", "NUnitTests\NUnitTests.csproj", "{429B8B5C-8483-4680-BC8D-E2792574F6A4}"
-EndProject
-Global
-	GlobalSection(TestCaseManagementSettings) = postSolution
-		CategoryFile = AdapterLab.vsmdi
-	EndGlobalSection
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Debug|Mixed Platforms = Debug|Mixed Platforms
-		Debug|x86 = Debug|x86
-		Release|Any CPU = Release|Any CPU
-		Release|Mixed Platforms = Release|Mixed Platforms
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{AD661DEE-4205-4289-9E77-E4F264809DE3}.Debug|Any CPU.ActiveCfg = Debug|x86
-		{AD661DEE-4205-4289-9E77-E4F264809DE3}.Debug|Mixed Platforms.ActiveCfg = Debug|x86
-		{AD661DEE-4205-4289-9E77-E4F264809DE3}.Debug|Mixed Platforms.Build.0 = Debug|x86
-		{AD661DEE-4205-4289-9E77-E4F264809DE3}.Debug|x86.ActiveCfg = Debug|x86
-		{AD661DEE-4205-4289-9E77-E4F264809DE3}.Debug|x86.Build.0 = Debug|x86
-		{AD661DEE-4205-4289-9E77-E4F264809DE3}.Release|Any CPU.ActiveCfg = Release|x86
-		{AD661DEE-4205-4289-9E77-E4F264809DE3}.Release|Mixed Platforms.ActiveCfg = Release|x86
-		{AD661DEE-4205-4289-9E77-E4F264809DE3}.Release|Mixed Platforms.Build.0 = Release|x86
-		{AD661DEE-4205-4289-9E77-E4F264809DE3}.Release|x86.ActiveCfg = Release|x86
-		{AD661DEE-4205-4289-9E77-E4F264809DE3}.Release|x86.Build.0 = Release|x86
-		{429B8B5C-8483-4680-BC8D-E2792574F6A4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{429B8B5C-8483-4680-BC8D-E2792574F6A4}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{429B8B5C-8483-4680-BC8D-E2792574F6A4}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
-		{429B8B5C-8483-4680-BC8D-E2792574F6A4}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
-		{429B8B5C-8483-4680-BC8D-E2792574F6A4}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{429B8B5C-8483-4680-BC8D-E2792574F6A4}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{429B8B5C-8483-4680-BC8D-E2792574F6A4}.Release|Any CPU.Build.0 = Release|Any CPU
-		{429B8B5C-8483-4680-BC8D-E2792574F6A4}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
-		{429B8B5C-8483-4680-BC8D-E2792574F6A4}.Release|Mixed Platforms.Build.0 = Release|Any CPU
-		{429B8B5C-8483-4680-BC8D-E2792574F6A4}.Release|x86.ActiveCfg = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-EndGlobal
