+++ src/MTCAsset.cs
@@ -14,7 +14,6 @@
  *    limitations under the License.
  */
 
-using System;
 using System.Xml;
 
 namespace MTConnect
